+++ fluent_contents/extensions/__init__.py
@@ -8,6 +8,8 @@ The API uses a registration system.
 While plugins can be easily detected via ``__subclasses__()``, the register approach is less magic and more explicit.
 Having to do an explicit register ensures future compatibility with other API's like reversion.
 """
+from .base_models import OrderableModel
+from .inlines import BasePluginInlineFormSet, PluginInline, TabularPluginInline, StackedPluginInline
 from .model_fields import PluginUrlField, PluginFileField, PluginImageField
 from .pluginbase import PluginContext, ContentPlugin
 from .pluginpool import plugin_pool, PluginNotFound, PluginAlreadyRegistered
@@ -17,7 +19,8 @@ from fluent_contents.models import ContentItem      # Allow plugins to pick ever
 
 __all__ = (
     'PluginContext', 'ContentPlugin',
-    'ContentItem',
+    'ContentItem', 'OrderableModel'
+    'BasePluginInlineFormSet', 'PluginInline', 'TabularPluginInline', 'StackedPluginInline',
     'PluginUrlField', 'PluginFileField', 'PluginImageField',
     'ContentItemForm',
     'plugin_pool',