+++ fluent_contents/plugins/linklist/migrations/0001_initial.py
@@ -1,0 +1,78 @@
+# -*- coding: utf-8 -*-
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        # Adding model 'LinkListItem'
+        db.create_table('contentitem_linklist_linklistitem', (
+            ('contentitem_ptr', self.gf('django.db.models.fields.related.OneToOneField')(to=orm['fluent_contents.ContentItem'], unique=True, primary_key=True)),
+            ('title', self.gf('django.db.models.fields.CharField')(max_length=200)),
+        ))
+        db.send_create_signal('linklist', ['LinkListItem'])
+
+        # Adding model 'Link'
+        db.create_table('linklist_link', (
+            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('order', self.gf('django.db.models.fields.PositiveIntegerField')(db_index=True, null=True, blank=True)),
+            ('linklist', self.gf('django.db.models.fields.related.ForeignKey')(related_name='links', to=orm['linklist.LinkListItem'])),
+            ('title', self.gf('django.db.models.fields.CharField')(max_length=200)),
+            ('url', self.gf('any_urlfield.models.fields.AnyUrlField')(max_length=300)),
+        ))
+        db.send_create_signal('linklist', ['Link'])
+
+
+    def backwards(self, orm):
+        # Deleting model 'LinkListItem'
+        db.delete_table('contentitem_linklist_linklistitem')
+
+        # Deleting model 'Link'
+        db.delete_table('linklist_link')
+
+
+    models = {
+        'contenttypes.contenttype': {
+            'Meta': {'ordering': "('name',)", 'unique_together': "(('app_label', 'model'),)", 'object_name': 'ContentType', 'db_table': "'django_content_type'"},
+            'app_label': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'model': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100'})
+        },
+        'fluent_contents.contentitem': {
+            'Meta': {'ordering': "('placeholder', 'sort_order')", 'object_name': 'ContentItem'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'parent_id': ('django.db.models.fields.IntegerField', [], {'null': 'True'}),
+            'parent_type': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['contenttypes.ContentType']"}),
+            'placeholder': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "'contentitems'", 'null': 'True', 'on_delete': 'models.SET_NULL', 'to': "orm['fluent_contents.Placeholder']"}),
+            'polymorphic_ctype': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "'polymorphic_fluent_contents.contentitem_set'", 'null': 'True', 'to': "orm['contenttypes.ContentType']"}),
+            'sort_order': ('django.db.models.fields.IntegerField', [], {'default': '1', 'db_index': 'True'})
+        },
+        'fluent_contents.placeholder': {
+            'Meta': {'unique_together': "(('parent_type', 'parent_id', 'slot'),)", 'object_name': 'Placeholder'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'parent_id': ('django.db.models.fields.IntegerField', [], {'null': 'True'}),
+            'parent_type': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['contenttypes.ContentType']", 'null': 'True', 'blank': 'True'}),
+            'role': ('django.db.models.fields.CharField', [], {'default': "'m'", 'max_length': '1'}),
+            'slot': ('django.db.models.fields.SlugField', [], {'max_length': '50'}),
+            'title': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'})
+        },
+        'linklist.link': {
+            'Meta': {'object_name': 'Link'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'linklist': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "'links'", 'to': "orm['linklist.LinkListItem']"}),
+            'order': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True', 'null': 'True', 'blank': 'True'}),
+            'title': ('django.db.models.fields.CharField', [], {'max_length': '200'}),
+            'url': ('any_urlfield.models.fields.AnyUrlField', [], {'max_length': '300'})
+        },
+        'linklist.linklistitem': {
+            'Meta': {'ordering': "('placeholder', 'sort_order')", 'object_name': 'LinkListItem', 'db_table': "'contentitem_linklist_linklistitem'", '_ormbases': ['fluent_contents.ContentItem']},
+            'contentitem_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': "orm['fluent_contents.ContentItem']", 'unique': 'True', 'primary_key': 'True'}),
+            'title': ('django.db.models.fields.CharField', [], {'max_length': '200'})
+        }
+    }
+
+    complete_apps = ['linklist']